server/requirements.txt
cffi==1.14.5
click==7.1.2
cryptography==3.4.7
Flask==1.1.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
greenlet==1.0.0
importlib-metadata==3.10.1
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
pycparser==2.20
PyJWT==1.7.1
six==1.15.0
SQLAlchemy==1.4.8
typing-extensions==3.7.4.3
uuid==1.30
Werkzeug==1.0.1
zipp==3.4.1
